+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased - ReleaseDate
 
+## 0.3.18 - 2022-04-21
+
 * Update Github actions
     * Use a container with GLIBC 2.17 to build linux-gnu binary
     * Codecov with cargo-tarpaulin